ci: use GitLab CI/CD & remove Github Actions
Removes both the normal workflows from Github Actions, as well as deleting the
Dependabot configuration. We'll need to bring the latter back at a later point
in time.

Creates a pipeline definition which runs for all branches, either in a branch or
merge request pipeline depending on whether merge requests exist for a given
branch.

Alters the versions used for Node.js to the LTS versions 12, 14 & 16.

Creates parallel pipelines for each of the versions of Node.js.

Runs one, shared linting job, only on merge requests and non-protected branches.
The idea behind that is that you never work on a protected branch, thus linting
is not necessary to run outside of merge requests. You might say the same of
tests, but seeing as protected branches should contain working code, it is far
more important that it works than whether or not pre-existing code does not
conform to the style guidelines used.